do not assert, warn user and skip offending record
this assert was caused by a default of 1 second for report interval,
which led to cases where delta time between records was under 0.5 sec,
but it was always positive number.
default has been increased to max(5 sec, duration/60) in fs-drift
in this commit we just skip a record if it's delta time is < 0.1 sec
and warn the user.