Disable SNAFU logging
This is a corner case needed during our log testing.
To accurately count the number of synthetic log messages
generated by the log generator pod, we want to disable the regular SNAFU
logging. This will make it harder to debug, but this is only something
a user optionally sets in the benchmark-operator CR. The reason we want
to do this is, we plan to deploy only the log-generator pods to a separate
namespace and forward those logs to a separate kafka topic and then we can
reliably read off the offset in the topic without reading each message to
verify that all log messages generated by the log generator pod made it to kafka.

Signed-off-by: Sai Sindhur Malleni <smalleni@redhat.com>