Move uperf CI files to ci dir
New proposed benchmark format has a single module for each benchmark,
meaning now we need a new place to hold CI files. This moves those
files into the CI dir. Helps with pythonic structure as well, as
now we are not storing data files in the Python module source code.