Remove TODO on adding mutex args in config
Easiest way to implement this functionality right now is to add
a check in the setup method of a benchmark. Not a real great way
to do this with argparse or configargparse without using
subparsers, which isn't quite a clean solution anyways. Might want
to allow for creation of another sub-group in the Config to help
document optionally dependent args, but IMHO it's NBH.