Pin docutils to less than 17
Currently sphinx is incompatible with docutils 17, however the
statistics module we import lists it as a required dependency.
pip-tools will pin docutils to 17.1, making dependency resolution
for our documentation impossible. The statistics module uses
docutils only for testing, meaning it should be save to pin
to a not up-to-date version.