Add logic to manifest step to handle missing images
If an image fails to build and does not get pushed to quay, then
it cannot be added to the manifest. However, the manifest step
tries to add it in anyways and fails, meaning the manifest will
only be updated if all architectures for a docker image succeed.
We want to allow for the manifest to be updated even if one
of the architectures fails, so that while we debug the new
image can be used.