Add summary output for uperf
Uperf spews raw per-second data to the screen. Instead, we should
provide a summary of the results.

The only issue here, is the user doesn't have the contextg of what the
test was, since it is in XML and read in via a file. This could be later
improved to read in the XML to understand the sort of test.